supported VC++9.0/x64
git-svn-id: http://svn.osdn.net/svnroot/hamigaki@1752 c04fc264-1e22-0410-9499-a186082fc69e

+++ hamigaki/trunk/hamigaki/coroutine/detail/fiber_context.hpp
@@ -1,6 +1,6 @@
 // fiber_context.hpp: Win32 fiber based context implementation
 
-// Copyright Takeshi Mouri 2006-2008.
+// Copyright Takeshi Mouri 2006-2009.
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
@@ -108,7 +108,8 @@ public:
     {
         if (stack_size == -1)
             stack_size = 0;
-        context_ = ::CreateFiber(stack_size, &trampoline<Functor>, &f);
+        context_ = ::CreateFiber(
+            static_cast<unsigned long>(stack_size), &trampoline<Functor>, &f);
         BOOST_ASSERT(context_ != 0);
     }
